Rename InMemoryAnyOfFilter to InMemoryInFilter
